# [Enhancement][Refactor] optimize array_contains_all/array_contains_seq function (backport #51701)