--- a/fe/fe-core/src/main/java/com/starrocks/catalog/FunctionSet.java
+++ b/fe/fe-core/src/main/java/com/starrocks/catalog/FunctionSet.java
@@ -319,6 +319,7 @@ public class FunctionSet {
     public static final String ARRAY_AVG = "array_avg";
     public static final String ARRAY_CONTAINS = "array_contains";
     public static final String ARRAY_CONTAINS_ALL = "array_contains_all";
+    public static final String ARRAY_CONTAINS_SEQ = "array_contains_seq";
     public static final String ARRAY_CUM_SUM = "array_cum_sum";
 
     public static final String ARRAY_JOIN = "array_join";
@@ -699,6 +700,7 @@ public class FunctionSet {
             .add(ARRAY_CONCAT)
             .add(ARRAY_SLICE)
             .add(ARRAY_CONTAINS)
+            .add(ARRAY_CONTAINS_ALL)
             .add(ARRAY_POSITION)
             .build();