--- a/test/sql/test_array_fn/R/test_array_contains
+++ b/test/sql/test_array_fn/R/test_array_contains
@@ -247,3 +247,209 @@ select array_position(v3, v2) from t;
 -- result:
 1
 -- !result
+-- name: test_array_contains_all_and_seq
+CREATE TABLE t (
+  k bigint(20) NOT NULL,
+  arr_0 array<bigint(20)> NOT NULL,
+  arr_1 array<bigint(20)>,
+  arr_2 array<bigint(20)>
+) ENGINE=OLAP
+DUPLICATE KEY(`k`)
+DISTRIBUTED BY RANDOM BUCKETS 1
+PROPERTIES (
+"replication_num" = "1"
+);
+-- result:
+-- !result
+insert into t values 
+(1, [1,2,3], [1,2], [1]),
+(2, [1,2,null], [1,null], [null]),
+(3, [1,2,null],[3],[3]),
+(4, [1,2,null], null, [1,2,null]),
+(5, [1,2,null], [1,2,null], null),
+(6, [1,2,3],[],[]),
+(7, [null,null], [null,null,null], [null,null]),
+(8, [1,1,1,1,1,2], [1,2], [1]),
+(9, [1,1,1,1,1,null,2],[1,null,2],[null,2]);
+-- result:
+-- !result
+select array_contains_all(arr_0, arr_1) from t order by k;
+-- result:
+1
+1
+0
+None
+1
+1
+1
+1
+1
+-- !result
+select array_contains_all(arr_1, arr_0) from t order by k;
+-- result:
+0
+0
+0
+None
+1
+0
+1
+1
+1
+-- !result
+select array_contains_all(arr_0, arr_2) from t order by k;
+-- result:
+1
+1
+0
+1
+None
+1
+1
+1
+1
+-- !result
+select array_contains_all(arr_2, arr_0) from t order by k;
+-- result:
+0
+0
+0
+1
+None
+0
+1
+0
+0
+-- !result
+select array_contains_all(arr_1, arr_2) from t order by k;
+-- result:
+1
+1
+1
+None
+None
+1
+1
+1
+1
+-- !result
+select array_contains_all(arr_2, arr_1) from t order by k;
+-- result:
+0
+0
+1
+None
+None
+1
+1
+0
+0
+-- !result
+select array_contains_all([1,2,3,4], arr_0) from t order by k;
+-- result:
+1
+0
+0
+0
+0
+1
+0
+1
+0
+-- !result
+select array_contains_all([1,2,3,4], arr_1) from t order by k;
+-- result:
+1
+0
+1
+None
+0
+1
+0
+1
+0
+-- !result
+select array_contains_all([1,2,3,4,null], arr_1) from t order by k;
+-- result:
+1
+1
+1
+None
+1
+1
+1
+1
+1
+-- !result
+select array_contains_all(arr_0, [1,null]) from t order by k;
+-- result:
+0
+1
+1
+1
+1
+0
+0
+0
+1
+-- !result
+select array_contains_all(arr_0, []) from t order by k;
+-- result:
+1
+1
+1
+1
+1
+1
+1
+1
+1
+-- !result
+select array_contains_all(null, arr_0) from t order by k;
+-- result:
+None
+None
+None
+None
+None
+None
+None
+None
+None
+-- !result
+select array_contains_all(arr_1, null) from t order by k;
+-- result:
+None
+None
+None
+None
+None
+None
+None
+None
+None
+-- !result
+set @arr0 = array_repeat("abcdefg", 10000);
+-- result:
+-- !result
+set @arr1 = array_repeat("abcdef", 100000);
+-- result:
+-- !result
+select array_contains_all(@arr0, @arr1);
+-- result:
+0
+-- !result
+set @arr0 = array_generate(10000);
+-- result:
+-- !result
+set @arr1 = array_generate(20000);
+-- result:
+-- !result
+select array_contains_all(@arr0, @arr1);
+-- result:
+0
+-- !result
+select array_contains_all(@arr1, @arr0);
+-- result:
+1
+-- !result