## Why I'm doing:

## What I'm doing:

Fixes #issue

backport #51701

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [x] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5

## Why I'm doing:

## What I'm doing:

optimize the performance of `array_contains_all` and `array_contains_seq`, and reduce their implementation complexity from `O(m*n)` to `O(m+n)`

basic idea
1. for array_contains_all, choose the side with less data to build a hash table to speed up the search
2. for array_contains_seq, it is essentially similar to the string search algorithm. We can use the idea of ​​the KMP algorithm to speed up the search.


## Test
Here I use two very simple queries to compare the performance
```sql
-- query 1
set @arr0 = array_generate(100000);
set @arr1 = array_generate(200000);
select array_contains_all(@arr1, @arr0);

-- query 2
set @arr0=array_append(array_repeat(1,100000),2);
set @arr1=array_append(array_repeat(1,50000),2);
select array_contains_seq(@arr0,@arr1);
```

|    | baseline  | after optimization |
|  ----  | ----  | ---- |
| query 1  | 1.66s | 0.05s | 
| query 2  | 1.08s |  0.02s |


## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [x] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [x] This is a backport pr